* commit 'f63624014576132b33eda063d1f4cb1547cb0b4f':
  Add RecordFree to the GarbageCollector interface

+++ runtime/gc/heap.h
@@ -358,7 +358,7 @@ class Heap {
 
   // Freed bytes can be negative in cases where we copy objects from a compacted space to a
   // free-list backed space.
-  void RecordFree(ssize_t freed_objects, ssize_t freed_bytes);
+  void RecordFree(uint64_t freed_objects, int64_t freed_bytes);
 
   // Must be called if a field of an Object in the heap changes, and before any GC safe-point.
   // The call is not needed if NULL is stored in the field.
